requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.5
django-oauth-toolkit==1.0.0
djangorestframework==3.6.4
drfdocs==0.0.11
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.4
psycopg2==2.7.3.1
pytz==2017.2
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
